+++ lopatina.sofia/T3/polygon.hpp
@@ -1,0 +1,22 @@
+#ifndef POLYGON_HPP
+#define POLYGON_HPP
+
+#include <iostream>
+#include <vector>
+namespace lopatina
+{
+  struct Point
+  {
+    int x, y;
+  };
+
+  struct Polygon
+  {
+    std::vector<Point> points;
+  };
+
+  std::istream & operator>>(std::istream & in, Polygon & dest);
+  std::ostream & operator<<(std::ostream & out, const Polygon & data);
+}
+
+#endif
